Allow loading the DCD and DCL files with parallel workers
Each chunk read from the DCD file can be independently inserted into
the ets table. This may speed things up in two ways:

* The cost of the binary_to_term call: reading the next log chunk can
  start immediately.
* Concurrently calling ets:insert on multiple chunks may also be
  faster than sequentially inserting them from the same
  process. (Although this depends heavily on the internals of ets
  locking.)

When it comes to DCL files, the order of operations is significant, so
there is less chance to speed up things with concurrent chunk
processing. However, the cost of the binary_to_term calls can be saved
in this scenario too. And it is also possible to preprocess the
operations in the chunk in parallel by grouping together deletes and
writes. This costs more CPU in total, but the strictly serial code
path (applying the preprocessed operations on the actual table) will
become faster.

New configuration parameters, no_dcd_workers and no_dcl_workers are
added to control the maximum number of parallel workers (per table
loader) in the DCD and DCL processing steps respectively.